+++ core-mvp-binding/sample/src/main/java/ru/surfstudio/android/mvp/binding/rx/sample/easyadapter/ui/common/recycler/animator/StandardItemAnimator.kt
@@ -16,7 +16,7 @@
 
 package ru.surfstudio.android.mvp.binding.rx.sample.easyadapter.ui.common.recycler.animator
 
-import ru.surfstudio.android.easyadapter.animator.BaseItemAnimator
+import ru.surfstudio.android.recycler.extension.animator.BaseItemAnimator
 
 /**
  * Origin <a href="http://google.com">https://github.com/MaksTuev/EasyAdapter/tree/master/sample/src/main/java/ru/surfstudio/easyadapter/sample</a>