+++ easyadapter/lib-easyadapter/src/main/java/ru/surfstudio/android/easyadapter/holder/BindableViewHolder.java
@@ -21,7 +21,6 @@ import android.view.ViewGroup;
 import androidx.annotation.LayoutRes;
 import androidx.recyclerview.widget.RecyclerView;
 
-import ru.surfstudio.android.easyadapter.AnimatableViewHolder;
 import ru.surfstudio.android.easyadapter.controller.BindableItemController;
 
 /**
@@ -33,7 +32,7 @@ import ru.surfstudio.android.easyadapter.controller.BindableItemController;
  *
  * @param <T> data type
  */
-public abstract class BindableViewHolder<T> extends BaseViewHolder implements AnimatableViewHolder {
+public abstract class BindableViewHolder<T> extends BaseViewHolder {
 
     public BindableViewHolder(ViewGroup parent, @LayoutRes int layoutRes) {
         super(parent, layoutRes);