+++ easyadapter/RELEASE_NOTES.md
@@ -28,6 +28,8 @@
     implement custom logic for showing more elements trigger
 * ANDDEP-1044 **NO BACKWARD COMPATIBILITY**
   * Remove animations methods from `BaseViewHolder` to interface `AnimatableViewHolder`
+  * Move `AnimatableViewHolder` and `BaseItemAnimator` to
+    `recycler-extension` module
 ##### Easyadapter Pagination
 * ANDDEP-1038 Add easyadapter-pagination module
 ## 1.0.0